Add no_std support
Based on @termoshtt's #15, without the intrusive file splitting.  I also
addressed the other review comments I had left there.  The limitations
of `no_std` are the missing implementations of `Error` and methods based
on `Float`, although `FloatCore` is now used for a few things.  Format
widths are also not supported, as I couldn't find a way to do this
without using a temporary `String`.

Co-authored-by: Toshiki Teramura <toshiki.teramura@gmail.com>