from_str_radix: Error if radix is > 18
Radix 19 and 20 introduce 'i' and 'j' as digits and this conflicts with
how complex are parsed. Before this fix, it could parse
`Num::from_str_radix("1ij + 1ij", 20)` in a pretty surprising way.

The `Num` docs already describe a rule for our situation - we might not
support every radix in our implementation; we can return an error in
this case instead.

A new (internal) error kind is introduced for this.

It is necessary to preserve the previous panicking behaviour for
radix > 36 (this is what primitive types do - but other types are
allowed to do differently, including never panicking).